Fixing different issues
git-tfs-id: [http://developmentvlan:8585/tfs/vanrise.collection]$/;C29600

+++ TOneV2/Code/TOneV2/TOne.WhS.BusinessEntity.Data.SQL/SupplierPriceListDataManager.cs
@@ -29,7 +29,7 @@ namespace TOne.WhS.BusinessEntity.Data.SQL
                 SupplierId = (int)reader["SupplierID"],
                 CurrencyId = (int)reader["CurrencyID"],
                 PriceListId = (int)reader["ID"],
-                FileId = (int)reader["FileID"]
+                FileId = (long)reader["FileID"]
             };
             return supplierPriceList;
         }
